requirements.txt
pyyaml
pycrypto
py2app